OpenAI Agents SDK/Multi-Agent Tutor System/pyproject.toml
[project]
name = "multi-agent-tutor-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "litellm>=1.67.5",
    "openai-agents>=0.0.14",
]
